[KGEN] Support fixed-width integers in parameter operators (#4515)
Treat signless integers as unsigned.

MODULAR_ORIG_COMMIT_REV_ID: dd1df2d5d5b9dee52fc75c744dac98008ca7c37d

+++ KGEN/test/kgen-dialect/kgen-errors.mlir
@@ -44,8 +44,8 @@ kgen.generator @test<p1>() {
 
 // -----
 
-// expected-error @+1 {{operator requires an index type}}
-%0 = kgen.param.constant: i32 = <shl(1, 2)>
+// expected-error @+1 {{operator requires an index or integer type}}
+%0 = kgen.param.constant: f32 = <shl(1., 2.)>
 
 // -----
